+++ CHANGELOG.md
@@ -4,4 +4,5 @@ FEATURES:
 
 * **New Data Source:** `clouddk_locations`
 * **New Data Source:** `clouddk_packages`
+* **New Data Source:** `clouddk_servers`
 * **New Data Source:** `clouddk_templates`